Fix various clippy warnings about unused code and variables
