+++ mountpoint-s3/tests/fuse_tests/mod.rs
@@ -15,11 +15,16 @@ mod write_test;
 use std::ffi::OsStr;
 use std::fs::ReadDir;
 
+use aws_sdk_s3::primitives::ByteStream;
+use aws_sdk_sts::config::Region;
 use fuser::{BackgroundSession, MountOption, Session};
+use futures::Future;
 use mountpoint_s3::fuse::S3FuseFilesystem;
 use mountpoint_s3::prefix::Prefix;
 use mountpoint_s3::S3FilesystemConfig;
 use mountpoint_s3_client::types::PutObjectParams;
+use rand::RngCore;
+use rand_chacha::rand_core::OsRng;
 use tempfile::TempDir;
 
 pub trait TestClient: Send {
@@ -189,8 +194,6 @@ mod mock_session {
 
 #[cfg(feature = "s3_tests")]
 mod s3_session {
-    use crate::common::{get_test_bucket_and_prefix, get_test_region};
-
     use super::*;
 
     use std::future::Future;
@@ -394,6 +397,7 @@ mod s3_session {
         }
     }
 }
+
 /// Take a `read_dir` iterator and return the entry names
 pub fn read_dir_to_entry_names(read_dir_iter: ReadDir) -> Vec<String> {
     read_dir_iter
@@ -407,4 +411,56 @@ pub fn read_dir_to_entry_names(read_dir_iter: ReadDir) -> Vec<String> {
             name.to_owned()
         })
         .collect::<Vec<_>>()
+}
+
+pub fn get_test_bucket_and_prefix(test_name: &str) -> (String, String) {
+    let bucket = std::env::var("S3_BUCKET_NAME").expect("Set S3_BUCKET_NAME to run integration tests");
+
+    // Generate a random nonce to make sure this prefix is truly unique
+    let nonce = OsRng.next_u64();
+
+    // Prefix always has a trailing "/" to keep meaning in sync with the S3 API.
+    let prefix = std::env::var("S3_BUCKET_TEST_PREFIX").unwrap_or(String::from("mountpoint-test/"));
+    assert!(prefix.ends_with('/'), "S3_BUCKET_TEST_PREFIX should end in '/'");
+
+    let prefix = format!("{prefix}{test_name}/{nonce}/");
+
+    (bucket, prefix)
+}
+
+pub fn get_test_bucket_forbidden() -> String {
+    std::env::var("S3_FORBIDDEN_BUCKET_NAME").expect("Set S3_FORBIDDEN_BUCKET_NAME to run integration tests")
+}
+
+pub fn get_test_region() -> String {
+    std::env::var("S3_REGION").expect("Set S3_REGION to run integration tests")
+}
+
+pub fn get_subsession_iam_role() -> String {
+    std::env::var("S3_SUBSESSION_IAM_ROLE").expect("Set S3_SUBSESSION_IAM_ROLE to run integration tests")
+}
+
+pub fn create_objects(bucket: &str, prefix: &str, region: &str, key: &str, value: &[u8]) {
+    let config = tokio_block_on(aws_config::from_env().region(Region::new(region.to_string())).load());
+    let sdk_client = aws_sdk_s3::Client::new(&config);
+    let full_key = format!("{prefix}{key}");
+    tokio_block_on(async move {
+        sdk_client
+            .put_object()
+            .bucket(bucket)
+            .key(full_key)
+            .body(ByteStream::from(value.to_vec()))
+            .send()
+            .await
+            .unwrap()
+    });
+}
+
+pub fn tokio_block_on<F: Future>(future: F) -> F::Output {
+    let runtime = tokio::runtime::Builder::new_current_thread()
+        .enable_io()
+        .enable_time()
+        .build()
+        .unwrap();
+    runtime.block_on(future)
 }
